Add GenMail.Core library, WPF UI, pipeline, dedupe stores and unit tests

+++ src/GenMail.Core/Models/Records.cs
@@ -23,7 +23,7 @@ public sealed record ProcessingCounters(
     long QualityRejected,
     long DuplicateSkipped,
     long EmailsWritten,
-    int OutputFilesCreated,
+    long OutputFilesCreated,
     int? RowsPerOutputFile);
 
 public sealed record ProcessingResult(
